rest/requirements.txt
django==2.2.7
coverage>=4.5.4
psycopg2-binary
gunicorn
tblib
djangorestframework==3.10.3
django-cors-headers
django-filter
tqdm
gensim
